+++ cli/create.h
@@ -3,16 +3,16 @@
 
 #include <rawstor.h>
 
-#include <stddef.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-int rawstor_cli_create_by_target(const char* target, size_t size);
+int rawstor_cli_create_by_target(const char* target, uint64_t size);
 
 int rawstor_cli_create_by_location(
-    const char* location, const char* uuid_string, size_t size
+    const char* location, const char* uuid_string, uint64_t size
 );
 
 #ifdef __cplusplus
