+++ cli/units.c
@@ -32,10 +32,10 @@ static int unit_to_shift(const char unit) {
     }
 }
 
-int rawstor_cli_size_to_bytes(const char* s, size_t* out) {
-    size_t value;
+int rawstor_cli_size_to_bytes(const char* s, uint64_t* out) {
+    uint64_t value;
     char unit;
-    if (sscanf(s, "%zu%c", &value, &unit) != 2) {
+    if (sscanf(s, "%llu%c", &value, &unit) != 2) {
         return -EINVAL;
     }
 
@@ -53,7 +53,7 @@ int rawstor_cli_size_to_bytes(const char* s, size_t* out) {
     return 0;
 }
 
-int rawstor_cli_bytes_to_size(size_t value, char* buf, size_t size) {
+int rawstor_cli_bytes_to_size(uint64_t value, char* buf, size_t size) {
     const char units[] = "BKMGTPE";
     size_t i;
     for (i = 0; i < sizeof(units) - 2; ++i) {
@@ -62,5 +62,5 @@ int rawstor_cli_bytes_to_size(size_t value, char* buf, size_t size) {
         }
         value >>= 10;
     }
-    return snprintf(buf, size, "%zu%c", value, units[i]);
+    return snprintf(buf, size, "%llu%c", value, units[i]);
 }
